fix saving logic, add favorite icon on details screen

+++ app/src/main/java/ru/justlearn/data/mappers/database/DefinitionEntityMapper.kt
@@ -5,7 +5,7 @@ import ru.justlearn.domain.Definition
 
 object DefinitionEntityMapper {
 
-    fun mapToEntity(meaningId: Int, definition: Definition): DefinitionEntity = with(definition) {
+    fun mapToEntity(meaningId: Long, definition: Definition): DefinitionEntity = with(definition) {
         DefinitionEntity(
             meaningId = meaningId,
             definition = this.definition,
