+++ app/src/main/java/ru/justlearn/data/saved/entities/AntonymEntity.kt
@@ -21,7 +21,7 @@ import androidx.room.PrimaryKey
 )
 data class AntonymEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val meaningId: Int,
+    val id: Long = 0,
+    val meaningId: Long,
     val value: String,
 )
