+++ app/src/main/java/ru/justlearn/data/saved/entities/MeaningEntity.kt
@@ -21,7 +21,7 @@ import androidx.room.PrimaryKey
 )
 data class MeaningEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
-    val wordId: Int,
+    val id: Long = 0,
+    val wordId: Long,
     val partOfSpeech: String,
 )
